+++ src/version_set.rs
@@ -68,6 +68,13 @@ pub trait VersionSet: Debug + Display + Clone + Eq {
             .complement()
     }
 
+    /// The set of versions that are in `self` but not in `other`.
+    ///
+    /// Implementations can override this to avoid materializing the complement.
+    fn difference(&self, other: &Self) -> Self {
+        self.intersection(&other.complement())
+    }
+
     /// Whether the ranges have no overlapping segments.
     fn is_disjoint(&self, other: &Self) -> bool {
         self.intersection(other) == Self::empty()
@@ -109,6 +116,10 @@ impl<T: Debug + Display + Clone + Eq + Ord> VersionSet for Ranges<T> {
 
     fn union(&self, other: &Self) -> Self {
         Ranges::union(self, other)
+    }
+
+    fn difference(&self, other: &Self) -> Self {
+        Ranges::difference(self, other)
     }
 
     fn is_disjoint(&self, other: &Self) -> bool {